// src/app/page/master/paket/paket.interface.ts
import {EventEmitter} from '@angular/core';

export interface Paket {
	
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	id?: any;
	kode?: any;
	version?: any;
	label?: any;
	code?: any;
	nama?: any;
	kodeEx?: any;
	namaEx?: any;
	value?: any;
	kdJenisPaket?: any;
	kdJenisTransaksi?: any;
 }